+++ cmd/frontend/db/authz_mock.go
@@ -2,8 +2,11 @@ package db
 
 import (
 	"context"
+
+	"github.com/sourcegraph/sourcegraph/cmd/frontend/types"
 )
 
 type MockAuthz struct {
 	GrantPendingPermissions func(ctx context.Context, args *GrantPendingPermissionsArgs) error
+	AuthorizedRepos         func(ctx context.Context, args *AuthorizedReposArgs) ([]*types.Repo, error)
 }