+++ enterprise/cmd/frontend/internal/authz/permissions.go
@@ -7,17 +7,6 @@ import (
 	otlog "github.com/opentracing/opentracing-go/log"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/authz"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/types"
-	"github.com/sourcegraph/sourcegraph/internal/api"
-	"github.com/sourcegraph/sourcegraph/internal/extsvc/bitbucketserver"
-)
-
-// ProviderType is the type of provider implementation for the permissions.
-type ProviderType string
-
-// The list of available provider types.
-const (
-	ProviderBitbucketServer ProviderType = bitbucketserver.ServiceType
-	ProviderSourcegraph     ProviderType = "sourcegraph"
 )
 
 // UserPermissions are the permissions of a user to perform an action
@@ -28,7 +17,7 @@ type UserPermissions struct {
 	Perm      authz.Perms
 	Type      authz.PermType
 	IDs       *roaring.Bitmap
-	Provider  ProviderType
+	Provider  authz.ProviderType
 	UpdatedAt time.Time
 }
 
@@ -40,13 +29,15 @@ func (p *UserPermissions) Expired(ttl time.Duration, now time.Time) bool {
 // AuthorizedRepos returns the intersection of the given repository IDs with
 // the authorized IDs.
 func (p *UserPermissions) AuthorizedRepos(repos []*types.Repo) []authz.RepoPerms {
-	if p.Type != authz.PermRepos {
-		return nil
+	// Return directly if it's used for wrong permissions type or no permissions available.
+	if p.Type != authz.PermRepos ||
+		p.IDs == nil || p.IDs.GetCardinality() == 0 {
+		return []authz.RepoPerms{}
 	}
 
 	perms := make([]authz.RepoPerms, 0, len(repos))
 	for _, r := range repos {
-		if r.ID != 0 && p.IDs != nil && p.IDs.Contains(uint32(r.ID)) {
+		if r.ID != 0 && p.IDs.Contains(uint32(r.ID)) {
 			perms = append(perms, authz.RepoPerms{Repo: r, Perms: p.Perm})
 		}
 	}
@@ -78,30 +69,13 @@ type RepoPermissions struct {
 	RepoID    int32
 	Perm      authz.Perms
 	UserIDs   *roaring.Bitmap
-	Provider  ProviderType
+	Provider  authz.ProviderType
 	UpdatedAt time.Time
 }
 
 // Expired returns true if these RepoPermissions have elapsed the given ttl.
 func (p *RepoPermissions) Expired(ttl time.Duration, now time.Time) bool {
 	return !now.Before(p.UpdatedAt.Add(ttl))
-}
-
-// AuthorizedUsers returns the intersection of the given user IDs with
-// the authorized IDs.
-func (p *RepoPermissions) AuthorizedUsers(users []*types.User) []authz.RepoPerms {
-	perms := make([]authz.RepoPerms, 0, len(users))
-	for _, u := range users {
-		if u.ID != 0 && p.UserIDs != nil && p.UserIDs.Contains(uint32(u.ID)) {
-			perms = append(perms, authz.RepoPerms{
-				Repo: &types.Repo{
-					ID: api.RepoID(p.RepoID),
-				},
-				Perms: p.Perm,
-			})
-		}
-	}
-	return perms
 }
 
 // TracingFields returns tracing fields for the opentracing log.
@@ -134,27 +108,6 @@ type UserPendingPermissions struct {
 	Type      authz.PermType
 	IDs       *roaring.Bitmap
 	UpdatedAt time.Time
-}
-
-// Expired returns true if these UserPendingPermissions have elapsed the given ttl.
-func (p *UserPendingPermissions) Expired(ttl time.Duration, now time.Time) bool {
-	return !now.Before(p.UpdatedAt.Add(ttl))
-}
-
-// AuthorizedRepos returns the intersection of the given repository IDs with
-// the authorized IDs.
-func (p *UserPendingPermissions) AuthorizedRepos(repos []*types.Repo) []authz.RepoPerms {
-	if p.Type != authz.PermRepos {
-		return nil
-	}
-
-	perms := make([]authz.RepoPerms, 0, len(repos))
-	for _, r := range repos {
-		if r.ID != 0 && p.IDs != nil && p.IDs.Contains(uint32(r.ID)) {
-			perms = append(perms, authz.RepoPerms{Repo: r, Perms: p.Perm})
-		}
-	}
-	return perms
 }
 
 // TracingFields returns tracing fields for the opentracing log.